Changed license from sloppy modifaction of zLib to C4L
I've implemented by five-term license, which is inspired partly by CCSA and partly by zLib. 

This is an unofficial summary of what the license does, and is not legally binding - 
Term 1 (I) - Disallows implementing closed source forks of this software. 
Term 2 (II) - Disallows forks from falsely claiming to be the work of others, including the original creator of this software. 
Term 3 (III) - Disallows non share-alike use of this software. 
Term 4 (V) - Disallows the fork from making malicious changes to the source code and then displaying itself as the original, unmodified work.